Add `racket-glossary` index entry
This makes it possible to find the `racket-glossary` documentation after
package installation with

  raco docs racket-glossary

Ideally, I'd like to have the browser opened with the glossary directly,
but having the title listed as search term so a user can click on it is
good enough for now.

Thanks to Sorawee for the heads up on `section-index`.