Add `combined-list-pict`
The algorithm is hacky in that it uses a module-global hash to store the
`pict`s from `cons-cell-pict` calls to reference them later. The hash
keys are individual list items, which means that having an item value
duplicated across `combined-list-pict` arguments may result in a wrong
result picture.

This hack may make the algorithm unsuitable for uses outside the simple
examples in the Racket glossary, which take the above limitation of the
code into account.

Also adjust some constants to make the generated `pict`s a bit nicer.