Add to-be stats script
I want to use this script to print some statistics about the glossary
entries so far, for example the ratio of written entries in each
category.

The current file contains parsing to `entry` structs and a few helper
functions to calculate further information from an `entry` (paragraphs,
paragraph count, and whether the entry is only a reference to another
entry).

The file uses some heuristics. It's not completely reliable, but it
should be enough to get an idea of the state of the glossary. In the
end, only careful reading by a human can determine if an entry can be
considered done.