Update README "Preview" section
The glossary is now public in rendered form on the Racket package
server.

I have embedded the current output of `glossary-stats.rkt`. At the
moment, the output is pasted into the README manually. Maybe I'll extend
the script in the future, so the script can update the README with a
command line option.