Change "Set" entry to "intermediate"
I had categorized this entry as "basic" since in some other languages
(at least Python), hash sets are a built-in data type. On the other
hand, although there are hash sets in Racket, not even a part of them is
in `racket/base`, so I consider sets as secondary to containers like
lists, vectors and maps.

Moreover, "set" without qualification is also a more generic term in
Racket, which includes lists, for example. Then you could argue why we'd
especially should describe hash sets but not other sets (as sets; there
are already entries on lists and vectors).