Use PNG images for "List" entry
Since the exported images cut off a part of the text box borders, I
added a _white_ continuous line under each of the images.

Also make the scale factor a bit smaller. I didn't notice it needed to
be smaller because the only image before (in the strings entry)
contained only lowercase text.