Make SVG images dark mode-compatible (WIP)
The SVG images exported from LibreOffice Draw by default are black on a
white backgroud. However, for an operating system or browser set to dark
mode, we want the opposite colors.

This script post-processes the SVG files to add style information and
set the `stroke` attribute from `rgb(0,0,0)` to `currentColor`.

The current code is incomplete. It contains the detection for whether a
conversion to dark mode support is needed, but the code doesn't contain
the actual conversion yet.