Fix SVG image for dark mode
The new code contains the updating for the style and color so the file
works for light and dark mode.

Contrary to what I had thought, it's necessary to change the `fill`
attribute in addition to the `stroke` attribute (and for maybe even more
attributes). Otherwise the text color remains black even in dark mode.

TODO:

- Write short-form tags (`<rect .../>` instead of `<rect ...></rect>`).
  There's a parameter for that.

- Iterate over all SVG files in the `scribblings` directory and fix only
  the files that need fixing.