Change no text to space
In `images.odg`, change the content of the right halves of the cons
cells from no text to a single space character. Somehow this seems to
fix LibreOffice Draw's confusion that led to a resizing of the right
halves.

This is in line with the observation that the right halves which contain
"()" were never resized, only empty squares were.