Change style of entry subsections
I would have preferred to add the style information to `subsubsection`,
but that doesn't seem to be possible at the moment. See
https://github.com/racket/scribble/issues/352 .

As a workaround, I use an `elem`, but then I have to apply several
styles from the `h5` element, which is used for `subsubsection`. I
looked up these styles in the Firefox developer tools. They come from
`scibble-manual.css`, so they shouldn't be OS or browser-specific. I'll
use `subsubsection` instead of `elem` when the above bug is fixed.

I also tried the approach with the extra CSS (see
https://docs.racket-lang.org/scribble/extra-style.html ) to remove the
border line from the `h5` element, but since my custom file is included
_before_ the default Scribble CSS files, the custom options are
overwritten.