Add new glossary entry function
Specify information on cross references and stubs explicitly, so we
won't need the current heuristics in `glossary-stats.rkt`.

At the moment, only the "Arity" entry is defined with `glossary-entry2`,
but the other entries are going to use the new function as well and the
old `glossary-entry` function will be replaced by the new one.