Settle on "Form" as the canonical entry
Both "form" and "macro" are used frequently and synonymously, but in
different situations. "Form" is usually used for the standard forms in
the language (say, `if` and `for/list`). "Macro" is usually used when
_defining_ new forms.

Therefore, I had considered having two entries, "Form" and "Macro" for
the different purposes, but I think that would be confusing. Instead,
I'd like to have a "general" part in the "Form" entry and a part about
defining new forms.