Don't typeset entry depending on `OMIT_STUB_ENTRIES`
If the environment variable `OMIT_STUB_ENTRIES` is set to "1", register
the glossary entry, but don't typeset it. That way, `glossary-stats.rkt`
can still display the missing entries.

Caveat: If an entry isn't typeset, all references to the stubs will be
broken and displayed accordingly as broken links in the HTML.

I decided to realize this feature with an environment variable. If it
was a flag in the source code, there would be a risk that I accidentally
commit the Scribble file with the flag set, which would cause the stub
entries to be omitted also on the package server. On the other hand,
the environment variable can only be set locally.