Add `glossary-inline-link`
This function can be used to create link to glossary entries from the
regular text, i.e. inline. Different from `secref`, with this function
you can specify the link text instead of having it fixed at the
(capitalized) section title, which is a bit obtrusive when used within a
sentence.